+++ CHANGELOG.md
@@ -3,9 +3,21 @@
 All notable changes to this project will be documented in this file. This project adheres to Semantic Versioning.
 
 
+## 2.1.0 - 2016-01-17
+
+### Added
+
+* `style` argument for elements `lin`,`rec`,`cir`,`arc`,`ply`.
+* `style` as first argument for `stroke`,`fill` and `drw`, optionally followed by a svg path definition string.
+
+### Changed
+
+* State stack reimplemented.
+
+
 ## 1.1.0 - 2016-01-08
 
 ### Added
 
-    CHANGELOG.md @goessner.
+  CHANGELOG.md @goessner.